pyproject.toml
[tool.poetry]
name = "api-reqres"
version = "0.1.0"
description = ""
authors = ["Alexander Santalov <vgtrk520@yandex.ru>"]
readme = "README.md"
packages = [{include = "api_reqres"}]

[tool.poetry.dependencies]
python = "^3.10"
selene = "2.0.0b17"
pytest-voluptuous = "1.2.0"
requests = "2.28.2"
pytest = "7.2.1"
voluptuous = "0.13.1"
python-dotenv = "0.21.0"
allure-pytest = "2.12.0"
curlify = "2.2.1"
faker = "^18.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = " -s --alluredir=allure-results --allure-no-capture"
